Update README.md
Config files for my GitHub profile. This project involved the analysis of two historical datasets (a clinical trial datasets and a dataset listing pharmaceutical violations) to extract valuable insights that can aid medical research and strategic decision making in the medical field. The steps were carried out in 3 phases on databricks using (PySpark(RDD and Dataframe) and Spark SQL).

The zipped file contains the steps used to import and pre-process the data. It also contains the different analysis done on databricks using (PySpark(RDD and Dataframe) and Spark SQL) as seperate files.